+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Added
 * Support for table unions in table builder
 * New column property `raise_missing_arg_exception` to allow for some spark function inputs to be missing
+* `add`, `sub`, `mul` and `div` spark functions
+
+### Breaking Change
+* Renamed `power` spark function to `scaled_power` to prevent conflict with native spark function 
 
 ## [0.1.8] - 2024-03-25
 ### Added